#23 - Added the validation class for checking the unique email address - Deleted unnecessary spaces - Added the new packages Email and Password and changed import files
