+++ src/main/java/com/gachigang/ontherun/common/validator/Email/UniqueEmail.java
@@ -1,4 +1,4 @@
-package com.gachigang.ontherun.common.validator;
+package com.gachigang.ontherun.common.validator.Email;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -6,7 +6,14 @@ import java.lang.annotation.*;
 
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-
+/**
+ * The @UniqueEmail annotation is used to verify that a given email has already existed in the Database.
+ * The annotation accepts a `message` attribute, which specifies the error
+ * message to be shown if the email address exists in the Database. It also supports
+ * the standard `groups` and `payload` attributes for use in validation
+ * groups and payloads.
+ * The annotation is used by the UniqueEmailValidator class to perform the actual validation.
+ */
 @Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE,
         ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE})
 @Retention(RUNTIME)
