+++ src/main/java/com/gachigang/ontherun/common/validator/Password/PasswordMatches.java
@@ -1,24 +1,22 @@
-package com.gachigang.ontherun.common.validator;
+package com.gachigang.ontherun.common.validator.Password;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
 import java.lang.annotation.*;
 
 /**
- * The @ValidPassword annotation is used to verify that a given string value is a valid password.
+ * The @PasswordMatches annotation is used to verify that password is equal to confirmPassword in the DTO.
  * The annotation accepts a `message` attribute, which specifies the error
- * message to be shown if the password is not valid. It also supports
+ * message to be shown if the password is not equals. It also supports
  * the standard `groups` and `payload` attributes for use in validation
  * groups and payloads.
  * The annotation is used by the PasswordMatcherValidator class to perform the actual validation.
  */
-
-@Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE,
-        ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE})
+@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @Constraint(validatedBy = PasswordMatchesValidator.class)
-public @interface ValidPassword {
+public @interface PasswordMatches {
 
     /**
      * Error message.
