+++ src/main/java/com/gachigang/ontherun/common/validator/Password/PasswordValidator.java
@@ -1,4 +1,4 @@
-package com.gachigang.ontherun.common.validator;
+package com.gachigang.ontherun.common.validator.Password;
 
 import com.gachigang.ontherun.common.ApplicationConstants;
 import lombok.NonNull;
