+++ src/main/java/com/gachigang/ontherun/common/validator/Password/ValidPassword.java
@@ -1,23 +1,23 @@
-package com.gachigang.ontherun.common.validator;
+package com.gachigang.ontherun.common.validator.Password;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
 import java.lang.annotation.*;
 
 /**
- * The @PasswordMatches annotation is used to verify that password is equal to confirmPassword in the DTO.
+ * The @ValidPassword annotation is used to verify that a given string value is a valid password.
  * The annotation accepts a `message` attribute, which specifies the error
- * message to be shown if the password is not equals. It also supports
+ * message to be shown if the password is not valid. It also supports
  * the standard `groups` and `payload` attributes for use in validation
  * groups and payloads.
  * The annotation is used by the PasswordMatcherValidator class to perform the actual validation.
  */
-
-@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
+@Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE,
+        ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @Constraint(validatedBy = PasswordMatchesValidator.class)
-public @interface PasswordMatches {
+public @interface ValidPassword {
 
     /**
      * Error message.
