+++ src/main/java/com/gachigang/ontherun/persistence/repository/UserRepository.java
@@ -5,6 +5,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
+/**
+ * Basic repository for the user.
+ */
 public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findUserByEmail(String email);
+    boolean existsByEmail(String email);
 }
