+++ src/test/java/com/gachigang/ontherun/common/validator/PasswordValidatorTest.java
@@ -1,5 +1,6 @@
 package com.gachigang.ontherun.common.validator;
 
+import com.gachigang.ontherun.common.validator.Password.PasswordValidator;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
